Update readme to clarify C support. (#209)
We will not support C++ headers, because C++ does not have a stable ABI like C (or use of extern "C" in C++ headers). Without a stable ABI it is impossible to know in what register/stack locations to pass arguments and this will lead to endless segmentation faults.

We could potentially support C++ headers which have extern "C" on every symbol, but we will not support C++ headers.